Fixes #7 - moves extract credentials to the runtime
* moves the methods for extracting credentials to runtime package
* adds tests for extract credentials.

This is better set up to be overriden in the future. This PR does not
allow for the ExtractCredentials to be overridden.
