+++ Makefile
@@ -1,4 +1,4 @@
-SOURCE_DIRS      = apb clients crd registries runtime
+SOURCE_DIRS      = bundle clients crd registries runtime
 SOURCES          := $(shell find . -name '*.go' -not -path "*/vendor/*" -not -path "*/.git/*")
 PACKAGES         := $(shell go list $(addprefix ./, $(addsuffix /... , $(shell ls -d */ | grep -v vendor))))
 COVERAGE_SVC     := travis-ci
